Move helper functions to shared file
